Fix makefile

+++ Makefile
@@ -101,6 +101,7 @@ clean:
 
 .PHONY: bump-version
 bump-version:
+	if [ git branch | grep \* | cut -f 2 -d \ != "master" ]; then echo "This must be ran from master"; exit 1; fi;
 	echo "$(BUMPED_VERSION)" > VERSION
 	git add VERSION
 	git commit -m "bump version $(VERSION) -> $(BUMPED_VERSION)"
